Queue file uploads instead of blocking on mutex
- add logging
- change how background process starts

+++ motion-notify.rb
@@ -1,31 +1,58 @@
 require 'getoptlong'
 require 'date'
+require 'fileutils'
 require 'drb'
 
 class Worker
-  def initialize(config_path, folder_id, unlink)
-    require 'google_drive'
+  def initialize(config_path, folder_id, unlink, log_path)
     @config_path = config_path
     @folder_id = folder_id
     @unlink = unlink
     @mutex = Mutex.new
+    @queue = []
+    if log_path
+      require 'logger'
+      @logger = Logger.new(log_path)
+    end
+    start_poll
   end
 
-  def upload(datetime, frame, picture_path)
+  def queue(datetime, frame, picture_path)
     @mutex.synchronize do
-      coll = parent
+      @queue.push([datetime, frame, picture_path])
+      log(:info, "Queued: #{datetime.to_s} (#{frame}), #{picture_path}")
+    end
+  end
 
-      # find or create folder for date
-      date_title = datetime.strftime("%Y-%m-%d")
-      sub_coll = coll.subcollection_by_title(date_title)
-      if sub_coll.nil?
-        sub_coll = coll.create_subcollection(date_title)
-      end
+  def start_poll
+    @poll ||= Thread.new do
+      loop do
+        @mutex.synchronize do
+          if !@queue.empty?
+            datetime, frame, picture_path = @queue.shift
+            log(:info, "Processing: #{datetime.to_s} (#{frame}), #{picture_path}")
+            begin
+              coll = parent
 
-      # upload picture
-      sub_coll.upload_from_file(picture_path, "#{datetime.strftime("%H:%M:%S")}-#{frame}")
-      if @unlink
-        File.unlink(picture_path)
+              # find or create folder for date
+              date_title = datetime.strftime("%Y-%m-%d")
+              sub_coll = coll.subcollection_by_title(date_title)
+              if sub_coll.nil?
+                sub_coll = coll.create_subcollection(date_title)
+              end
+
+              # upload picture
+              sub_coll.upload_from_file(picture_path, "#{datetime.strftime("%H:%M:%S")}-#{frame}")
+              if @unlink
+                File.unlink(picture_path)
+              end
+              log(:info, "Finished: #{datetime.to_s} (#{frame}), #{picture_path}")
+            rescue Exception => exp
+              log(:fatal, exp)
+            end
+          end
+        end
+        sleep 1
       end
     end
   end
@@ -33,11 +60,26 @@ class Worker
   private
 
   def session
-    @session ||= GoogleDrive::Session.from_config(@config_path)
+    if @session.nil?
+      require 'google_drive'
+      @session = GoogleDrive::Session.from_config(@config_path)
+    end
+    @session
   end
 
   def parent
     @parent ||= session.file_by_id(@folder_id)
+  end
+
+  def log(severity, msg)
+    if @logger
+      severity =
+        case severity
+        when :info  then Logger::INFO
+        when :fatal then Logger::FATAL
+        end
+      @logger.log(severity, msg)
+    end
   end
 end
 
@@ -65,6 +107,18 @@ def print_usage(program_name)
 
 -D drb-uri, --drb drb-uri:
    URI of DRb process
+
+-U, --no-unlink:
+   Disable unlinking of files
+
+-s, --start:
+   Start background job
+
+-P pid-path, --pid pid-path:
+   Path of PID file
+
+-l log-path, --log log-path:
+   Path of log file
 EOF
 end
 
@@ -77,6 +131,9 @@ opts = GetoptLong.new(
   [ '--picture',    '-p',  GetoptLong::REQUIRED_ARGUMENT ],
   [ '--drb',        '-D',  GetoptLong::REQUIRED_ARGUMENT ],
   [ '--no-unlink',  '-U',  GetoptLong::NO_ARGUMENT ],
+  [ '--start',      '-s',  GetoptLong::NO_ARGUMENT ],
+  [ '--pid',        '-P',  GetoptLong::REQUIRED_ARGUMENT ],
+  [ '--log',        '-l',  GetoptLong::REQUIRED_ARGUMENT ],
 )
 
 program_name = $0
@@ -87,6 +144,9 @@ frame = nil
 picture_path = nil
 unlink = true
 drb_uri = "druby://localhost:8787"
+start_background = false
+pid_path = nil
+log_path = nil
 
 opts.each do |opt, arg|
   case opt
@@ -107,6 +167,12 @@ opts.each do |opt, arg|
     drb_uri = arg
   when '--no-unlink'
     unlink = false
+  when '--start'
+    start_background = true
+  when '--pid'
+    pid_path = arg
+  when '--log'
+    log_path = arg
   end
 end
 
@@ -119,24 +185,30 @@ if config_path.nil?
   puts "--config is required"
   valid = false
 end
-if datetime.nil?
-  puts "--datetime is required"
+if pid_path.nil?
+  puts "--pid is required"
   valid = false
-else
-  begin
-    datetime = DateTime.strptime(datetime, "%Y-%m-%d %H:%M:%S")
-  rescue ArgumentError
-    puts "#{datetime} is not a valid datetime"
+end
+if !start_background
+  if datetime.nil?
+    puts "--datetime is required"
+    valid = false
+  else
+    begin
+      datetime = DateTime.strptime(datetime, "%Y-%m-%d %H:%M:%S")
+    rescue ArgumentError
+      puts "#{datetime} is not a valid datetime"
+      valid = false
+    end
+  end
+  if frame.nil?
+    puts "--frame is required"
     valid = false
   end
-end
-if frame.nil?
-  puts "--frame is required"
-  valid = false
-end
-if picture_path.nil?
-  puts "--picture is required"
-  valid = false
+  if picture_path.nil?
+    puts "--picture is required"
+    valid = false
+  end
 end
 
 if !valid
@@ -144,22 +216,43 @@ if !valid
   exit
 end
 
-# try to connect to existing DRb process first
-server = DRbObject.new_with_uri(drb_uri)
-tried = false
-begin
-  server.upload(datetime, frame, picture_path)
-rescue DRb::DRbConnError => exp
-  # failed, boot DRb
-  fork do
-    DRb.start_service(drb_uri, Worker.new(config_path, folder_id, unlink))
-    DRb.thread.join
-  end
-  sleep 5
-  if !tried
-    tried = true
+if start_background
+  at_exit { File.unlink(pid_path) }
+  DRb.start_service(drb_uri, Worker.new(config_path, folder_id, unlink, log_path))
+  DRb.thread.join
+else
+  # try to connect to existing DRb process first
+  server = DRbObject.new_with_uri(drb_uri)
+  tried = 0
+  begin
+    server.queue(datetime, frame, picture_path)
+  rescue DRb::DRbConnError => exp
+    if tried >= 5
+      raise "Tried 5 times to connect to background process, but failed"
+    end
+
+    # failed, boot DRb if pid file doesn't exist
+    if !File.exist?(pid_path)
+      FileUtils.touch(pid_path)
+      args = [
+        program_name, "--start",
+        "--config", config_path,
+        "--folder-id", folder_id,
+        "--pid", pid_path
+      ]
+      if log_path
+        args << "--log"
+        args << log_path
+      end
+      if !unlink
+        args << "--no-unlink"
+      end
+      pid = spawn(RbConfig.ruby, *args)
+      File.open(pid_path, 'w') { |f| f.puts(pid) }
+    end
+
+    sleep 1
+    tried += 1
     retry
-  else
-    raise "Couldn't start server"
   end
 end
